adaq7980: Software configurable trigger
The previous version of the SPI Engine framework used a fixed value
for the pulse period of the trigger and for the load configuration.

In order to make these parameters software configurable, two registers
were added in the register map of the axi: pulse_gen_period and
pulse_gen_load. The pulse width value remains constant, driven by
trigger_gen_pulse_width.